+++ ocaml-lsp-server/test/e2e-new/start_stop.ml
@@ -168,7 +168,7 @@ let%expect_test "start/stop" =
         "definitionProvider": true,
         "documentFormattingProvider": true,
         "documentHighlightProvider": true,
-        "documentLinkProvider": { "resolveProvider": false },
+        "documentLinkProvider": { "resolveProvider": true },
         "documentRangeFormattingProvider": true,
         "documentSymbolProvider": true,
         "executeCommandProvider": {